+++ src/components/Templates/Listings.js
@@ -10,7 +10,8 @@ import health from '@fortawesome/fontawesome-free-solid/faPlusSquare';
 import activities from '@fortawesome/fontawesome-free-solid/faFutbol';
 import food from '@fortawesome/fontawesome-free-solid/faCoffee';
 import wellbeing from '@fortawesome/fontawesome-free-solid/faLeaf';
-import axios from 'axios';
+// import axios from 'axios';
+import {loadData, mergeData} from '../../actions/index';
 // import { connect } from 'react-redux';
 // import * as actionCreators from '../actions/index';
 
@@ -26,65 +27,13 @@ class Listings extends React.Component {
       page: window.location.href.split('/').slice(-1)[0],
       results: []
     };
-
-    this.loadFilters = this.loadFilters.bind(this);
   }
 
-  getCategory(cat) {
-    // Basic Needs
-    // Family / Whanau Services
-    // Health
-    // Child Care
-    // Parents and Caregivers
-    switch(cat) {
-    case 'home':
-      return 'Family / Whanau Services, Child Care, Parents and Caregivers';
-    case 'health':
-      return 'Health';
-    case 'food':
-      return '';
-    case 'activities':
-      return '';
-    case 'wellbeing':
-      return 'Basic Needs';
-    default:
-      return '';
-    }
-  }
-
-  loadFilters(){
-    const GLOBAL_FILTER = 'community services card';
-    const RESOURCE_ID = process.env.REACT_APP_API_RESOURCE_ID;
-    const API_PATH = process.env.REACT_APP_API_PATH;
-    const CATEGORY = this.getCategory(this.state.page);
-    let fields = '*';
-
-    let where = `WHERE "SERVICE_DETAIL" LIKE '%${GLOBAL_FILTER}%'
-    AND "LEVEL_1_CATEGORY" = '${CATEGORY}'
-    `;
-    let whereArray = `WHERE "SERVICE_DETAIL" LIKE '%${GLOBAL_FILTER}%'
-    &&|| "LEVEL_1_CATEGORY" = '${CATEGORY.split(',')[0]}'
-    &&|| "LEVEL_1_CATEGORY" = '${CATEGORY.split(',')[1]}'
-    &&|| "LEVEL_1_CATEGORY" = '${CATEGORY.split(',')[2]}'
-    `;
-    // alert(CATEGORY.split(',')[0])
-
-    let sql =`SELECT ${fields} FROM "${RESOURCE_ID}" ${CATEGORY.length > 1 ? whereArray : where}`;
-    sql =  encodeURI(sql);
-    let url = `${API_PATH}datastore_search_sql?sql=${sql}`;
-    return axios.get(url).then((response)=>{
-      this.setState({results: response.data.result.records});
-    });
-  }
-    // let fields = '"LEVEL_1_CATEGORY" as name, COUNT(*) as num';
-    // let where = `WHERE "SERVICE_DETAIL" LIKE '%${GLOBAL_FILTER}%'
-    //   OR "SERVICE_TARGET_AUDIENCES" LIKE '%${GLOBAL_FILTER}%'
-    //   OR "COST_DESCRIPTION" LIKE '%${GLOBAL_FILTER}%'
-    //   OR "DELIVERY_METHODS" LIKE '%${GLOBAL_FILTER}%'`;
-
-    // let sql =`SELECT ${fields} FROM "${RESOURCE_ID}" ${where} GROUP BY name ORDER BY name`;
   componentDidMount() {
-    return this.loadFilters();
+    // return loadData(this.state.page);
+    // console.log('in listings', loadData(this.state.page))
+    this.setState({results: loadData(this.state.page)})
+    console.log(this.state.page)
   }
 
   renderTheme(state) {
@@ -132,24 +81,6 @@ class Listings extends React.Component {
     return unique;
   }
 
-  mergeData(services, results) {
-    const resArr = [];
-    const result = services.map((el, i) => {
-      var o = Object.assign({}, el);
-      o.FSD_ID = `0000${i+1}`;
-      return o;
-    });
-    Array.prototype.push.apply(result, results);
-    result.filter(function(item){
-      const i = resArr.findIndex(x => x.FSD_ID === item.FSD_ID);
-      if(i <= -1){
-        resArr.push({SERVICE_NAME: item.SERVICE_NAME, PROVIDER_NAME: item.PROVIDER_NAME, FSD_ID: item.FSD_ID, SERVICE_DETAIL: item.SERVICE_DETAIL});
-      }
-      return null;
-    });
-
-    return resArr;
-  }
   render(){
     document.querySelector('body').setAttribute('class',`${this.renderTheme(this.state.page)}-bg`);
     return (
@@ -161,20 +92,17 @@ class Listings extends React.Component {
         </header>
         <div className="container">
           <ul className="list-stripped">
-            {
-              this.mergeData(services, this.state.results).map((item, key) => {
+            {mergeData(services, this.state.results).map((item, key) => {
                 return <li key={key} className={`${this.renderTheme(this.state.page)}`}>
                   <a href={`#/${this.state.page}/${item.FSD_ID}`}>
                     <Image src="http://placekitten.com/200/300" alt="kitten" />
                     <span className="listing-details">
-                      <h3>{item.SERVICE_NAME}</h3>
-                      <h4 style={{color: 'white'}}>{item.SERVICE_PROVIDER}</h4>
+                      <h3>{item.SERVICE_NAME} - {item.PROVIDER_NAME}</h3>
                       <p>{this.text_truncate(item.SERVICE_DETAIL)}</p>
                     </span>
                   </a>
                 </li>;
-              })
-            }
+              })}
           </ul>
         </div>
       </div>
